Add controller
