+++ config/routes.rb
@@ -1,4 +1,9 @@
 Rails.application.routes.draw do
-  resources :artists
-  # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
+
+  namespace :api do
+    namespace :v1 do
+      resources :artists
+    end
+  end
+
 end
